+++ Gruntfile.js
@@ -3,7 +3,8 @@ module.exports = function(grunt) {
   require('load-grunt-tasks')(grunt);
 
   const config = {
-    src: ['Gruntfile.js', 'src/**/*.js'],
+    // Source files to JSHint and copy over
+    src: ['Gruntfile.js', 'src/**/*.js', 'test/**/*.test.js'],
 
     // In game > Script, bottom of it, click on 'Open local folder', copy and paste in here as 'dest'
     dest: '/Users/carlorizzante/Library/Application Support/Screeps/scripts/127_0_0_1___21025/default',
@@ -57,7 +58,7 @@ module.exports = function(grunt) {
     },
     watch: {
       files: config.src,
-      tasks: ['jshint']
+      tasks: ['jshint', 'mochaTest']
     },
     wait: {
       options: {
@@ -73,9 +74,23 @@ module.exports = function(grunt) {
           }
         }
       }
+    },
+    mochaTest: {
+      test: {
+        options: {
+          reporter: 'spec',
+          // captureFile: 'results.txt', // Optionally capture the reporter output to a file
+          quiet: false, // Optionally suppress output to standard out (defaults to false)
+          clearRequireCache: false, // Optionally clear the require cache before running tests (defaults to false)
+          clearCacheFilter: (key) => true, // Optionally defines which files should keep in cache
+          noFail: false // Optionally set to not fail on failed tests (will still fail on other errors)
+        },
+        src: ['test/**/*.test.js']
+      }
     }
   });
 
-  grunt.registerTask('default', ['jshint']);
+  grunt.registerTask('default', ['jshint', 'mochaTest']);
+  grunt.registerTask('test', ['mochaTest']);
   grunt.registerTask('update', ['clean', 'wait', 'copy', 'wait', 'sync']);
 }