+++ src/roles/role.builder.js
@@ -8,6 +8,7 @@ const rechargeTowers     = true
 
 module.exports = {
   run: creep => {
+    "use strict";
 
     const room = creep.room.name;
     const homeroom = creep.memory.homeroom;