+++ src/roles/role.harvester.js
@@ -12,6 +12,7 @@ const useSources         = true;
 
 module.exports = {
   run: creep => {
+    "use strict";
 
     const room = creep.room.name;
     const homeroom = creep.remember('homeroom');